Find a p-value given a certain number of observations in favor of the working hypothesis among a total number of observations when there are two types of observations

When an observation can either support a working theory or a rival theory, this function returns the p-value summarizing the evidence against the rival theory provided by the working theory.

Usage

find_p_two_types(
  obs_support,
  total_obs,
  odds = 1,
  weights = NULL,
  interpretation = FALSE
)

Arguments

obs_support

An integer representing the number of observations in favor of the working hypothesis. Must be less than or equal to the total.

total_obs

An integer representing the total number of observations made

odds

The odds of observing a rival versus working-theory observation. This can be interpreted as "bias" in observation. Or "relative ease" of observation.

interpretation

TRUE if the function returns text helping to interpret the result, FALSE (default option) to returns only the p-value

smoking_gun_weight

Double. Additional weight given to a single observation out of the total.

Value

Either a p-value (numeric, scalar) or a list containing the p-value and text containing an interpretation

Details

This function accomodates urns where working theory supporting observations are systematically easier or more difficulty to observe that rival supporting observations (odds)

And it currently also supports designs where one observation out of the total supporting the working theory is particularly compelling (a "smoking gun"). For example, if observation 1 is worth 2 other observations, the urn can reflect this by increasing the number of observations supporting the rival theory in the urn, accounting for the weight of that smoking gun observation.