Find odds given ideas about unequally easy evidence

Find odds

Usage

sens_analysis(obs_support, obs_oppose, total_obs, thep = 0.05)

Arguments

obs_support

An integer representing the number of observations in favor of the working hypothesis. Must be less than or equal to the total.

total_obs

An integer representing the total number of observations

thep

The p-value threshold

Value

The odds required for the p-value to be equal to that specified in thep

Details

A function that allows the researcher to input the likelihood of data or to find how mush bias in datacollection would be nescessary to obtain a p>0.05 or p>0.10 result using the same data.